+++ tests/testthat/test-add_p.R
@@ -1,4 +1,4 @@
-context("test-add_p")
+context("test-add_p.tbl_summary")
 
 test_that("add_p creates output without error/warning", {
   expect_error(
@@ -51,6 +51,29 @@ test_that("add_p defaults to clustered data with `group=` arg", {
   )
 })
 
+
+context("test-add_p.tbl_cross")
+
+test_that("add_p.tbl_cross", {
+  expect_error(
+    trial %>% tbl_cross(response, death) %>% add_p(),
+    NA
+  )
+  expect_error(
+    trial[c("trt", "grade")] %>% tbl_cross() %>% add_p(),
+    NA
+  )
+  expect_error(
+    trial[c("trt", "grade")] %>% tbl_cross() %>% add_p(source_note = TRUE),
+    NA
+  )
+  expect_error(
+    mtcars %>%
+      tbl_cross(gear, carb) %>%
+      add_p(test = "fisher.test"),
+    NA
+  )
+})